pythonProject — копия/requirements.txt
Django==2.2
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2

pip~=22.0.4
cryptography~=36.0.2
Jinja2~=3.1.1
setuptools~=62.1.0
coreapi~=2.3.3
uritemplate~=4.1.1
coreschema~=0.0.4
requests~=2.27.1
djangorestframework-jwt~=1.11.0